shobitparsing/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic>=2.0
pydantic-settings>=2.0
tableauhyperapi
pandas
requests
azure-storage-blob
pydantic-settings